Migrated MessageCreate to the BotEvent structure
- Migrated MessageCreate to the BotEvent structure
- Refactored some of the message create mapping code
- Removed deprecated mappings.json support
