+++ .gitignore
@@ -1,5 +1,4 @@
 build/
-mappings.json
 revcord.sqlite*
 
 # Editors
