+++ app/discord.ts
@@ -38,7 +38,7 @@ import UniversalExecutor from "./universalExecutor";
  * @param ping ID of the user to ping
  * @returns Formatted string
  */
-function formatMessage(
+export function formatMessage(
   attachments: Collection<string, Attachment>,
   content: string,
   mentions: MessageMentions,
@@ -128,188 +128,6 @@ function formatMessage(
   messageString = messageString.replace(TrailingNewlines, '');
 
   return messageString;
-}
-
-/**
- * Find a relevant mapping and direct a Discord message to Revolt
- * @param revolt Revolt client
- * @param discord Discord client
- * @param message Discord message object
- */
-export async function handleDiscordMessage(
-  revolt: RevoltClient,
-  discord: DiscordClient,
-  message: Message
-) {
-  try {
-    // Find target Revolt channel
-    const target = Main.mappings.find((mapping) => mapping.discord === message.channelId);
-
-    // Bot check
-    if (
-      target &&
-      message.applicationId !== discord.user.id &&
-      (!message.author.bot || target.allowBots)
-    ) {
-      // Prepare masquerade
-      const mask = {
-        // Support for new username system
-        name: truncate(
-          message.author.username +
-            (message.author.discriminator.length === 1
-              ? ""
-              : "#" + message.author.discriminator),
-          32
-        ),
-        avatar: message.author.avatarURL(),
-      };
-
-      // Handle replies
-      const reference = message.reference;
-      let replyPing: string;
-
-      let replyEmbed: ReplyObject;
-
-      if (reference) {
-        // Find cross-platform replies
-        const crossPlatformReference = Main.revoltCache.find(
-          (cached) => cached.createdMessage === reference.messageId
-        );
-
-        if (crossPlatformReference) {
-          replyPing = crossPlatformReference.parentMessage;
-        } else {
-          // Find same-platform replies
-          const samePlatformReference = Main.discordCache.find(
-            (cached) => cached.parentMessage === reference.messageId
-          );
-
-          if (samePlatformReference) {
-            replyPing = samePlatformReference.createdMessage;
-          } else {
-            // Fallback - this happens when someone replies to a message
-            // that was sent before the bot was started
-
-            // Wrap in another try-catch since it may fail
-            // if the bot doesn't have permission to view message history
-            try {
-              // Fetch referenced message
-              const sourceChannel = await discord.channels.fetch(
-                message.reference.channelId
-              );
-
-              if (sourceChannel instanceof TextChannel) {
-                const referenced = await sourceChannel.messages.fetch(
-                  message.reference.messageId
-                );
-
-                // Prepare reply embed
-                const formattedContent = formatMessage(
-                  referenced.attachments,
-                  referenced.content,
-                  referenced.mentions
-                );
-
-                replyEmbed = {
-                  pingable: false,
-                  entity:
-                    referenced.author.username + "#" + referenced.author.discriminator,
-                  entityImage: referenced.author.avatarURL(),
-                  content: formattedContent,
-                  attachments: [],
-                };
-
-                if (referenced.attachments.first()) {
-                  replyEmbed.attachments.push("file");
-                  replyEmbed.previewAttachment = referenced.attachments.first().url;
-                }
-              }
-            } catch (e) {
-              npmlog.warn("Discord", 'Bot lacks the "View message history" permission.');
-              npmlog.warn("Discord", e);
-            }
-          }
-        }
-      }
-
-      // Sticker
-      const sticker = message.stickers.first();
-      let stickerUrl = sticker && sticker.url;
-
-      // Format message content (parse emojis, mentions, images etc.)
-      const messageString = formatMessage(
-        message.attachments,
-        message.content,
-        message.mentions,
-        stickerUrl
-      );
-
-      // Prepare message object
-      // revolt.js doesn't support masquerade yet, but we can use them using this messy trick.
-      const messageObject = {
-        content: truncate(messageString, 1984),
-        masquerade: mask,
-        replies: replyPing
-          ? [
-              {
-                id: replyPing,
-                mention: false,
-              },
-            ]
-          : [],
-      } as any;
-
-      if (replyEmbed) {
-        if (typeof messageObject.embeds === "undefined") messageObject.embeds = [];
-        messageObject.embeds.push({
-          type: "Text",
-          icon_url: replyEmbed.entityImage,
-          title: replyEmbed.entity,
-          description: `**Reply to**: ${replyEmbed.content}`,
-        });
-      }
-
-      // Translate embeds, if present.
-      // Allow embeds only from bots, since a regular user
-      // shouldn't be able to send them.
-      if (message.embeds.length && message.author.bot) {
-        // Add an empty array
-        if (typeof messageObject.embeds === "undefined") messageObject.embeds = [];
-        
-        // Translate embed
-        try {
-          const embed = new RevcordEmbed().fromDiscord(message.embeds[0]).toRevolt();
-          
-          messageObject.embeds.push(embed);
-        } catch (e) {
-          npmlog.warn("Discord", "Failed to translate embed.");
-          npmlog.warn("Discord", e);
-        }
-      }
-
-      const sentMessage = await revolt.channels
-        .get(target.revolt)
-        .sendMessage(messageObject);
-
-      // Save in cache
-      Main.discordCache.push({
-        parentMessage: message.id,
-        parentAuthor: message.author.id,
-        createdMessage: sentMessage._id,
-        channelId: target.discord,
-      });
-    }
-  } catch (e) {
-    npmlog.warn("Revolt", "Couldn't send a message to Revolt");
-    npmlog.warn("Revolt", e);
-
-    if ("response" in e && "status" in e.response && e.response.status === 403) {
-      npmlog.error(
-        "Revolt",
-        "It seems the bot doesn't have enough permissions (most likely Masquerade)"
-      );
-    }
-  }
 }
 
 /**
